Added rest of the mocks and first running test.

+++ sparkles-core/src/test/scala/pfvalter/sparkles/core/framework/schemas/MockInput.scala
@@ -2,5 +2,5 @@ package pfvalter.sparkles.core.framework.schemas
 
 case class MockInput(
                            field1: String,
-                           field2: Int
+                           field2: Long
                          ) extends Product
