+++ sparkles-core/src/test/scala/pfvalter/sparkles/core/framework/schemas/MockOutput.scala
@@ -1,7 +1,7 @@
 package pfvalter.sparkles.core.framework.schemas
 
 case class MockOutput(
-                       fieldA: Int,
+                       fieldA: Long,
                        fieldB: String,
                        fieldC: Option[String]
                      ) extends Product
